Выполнил этап 2 ФЗ

+++ main-service/src/main/java/ru/practicum/service/CategoryServiceImpl.java
@@ -6,7 +6,6 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.category.NewCategoryDto;
 import ru.practicum.exception.AlreadyExistsException;
@@ -14,6 +13,7 @@ import ru.practicum.exception.NotFoundException;
 import ru.practicum.mapper.CategoryMapper;
 import ru.practicum.model.Category;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
 
 import java.util.List;
 
@@ -21,11 +21,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
     public CategoryDto saveCategory(NewCategoryDto newCategoryDto) {
-        if (categoryRepository.findByName(newCategoryDto.getName()).isPresent()) {
+        if (categoryRepository.existsByName(newCategoryDto.getName())) {
             throw new AlreadyExistsException("Категория с называнием  " + newCategoryDto.getName() + " существует");
         }
         return CategoryMapper.mapToCategoryDto(categoryRepository.save(CategoryMapper.mapToCategory(newCategoryDto)));
@@ -33,15 +34,13 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional
-    public CategoryDto patchCategory(Long id, CategoryDto categoryDto) {
+    public CategoryDto patchCategory(Long id, NewCategoryDto categoryDto) {
         Category categoryForPatch = categoryRepository.findById(id)
-                .orElseThrow(()-> new NotFoundException("Категории с id=" + id + " не существует"));
-        if (StringUtils.hasText(categoryDto.getName())) {
-            if (categoryRepository.findByName(categoryDto.getName()).isPresent()) {
-                throw new AlreadyExistsException("Категория с называнием  " + categoryDto.getName() + " существует");
-            }
-            categoryForPatch.setName(categoryDto.getName());
+                .orElseThrow(() -> new NotFoundException("Категории с id=" + id + " не существует"));
+        if (categoryRepository.existsByName(categoryDto.getName()) && !categoryDto.getName().equals(categoryForPatch.getName())) {
+            throw new AlreadyExistsException("Категория с называнием  " + categoryDto.getName() + " существует");
         }
+        categoryForPatch.setName(categoryDto.getName());
         return CategoryMapper.mapToCategoryDto(categoryRepository.save(categoryForPatch));
     }
 
@@ -51,7 +50,9 @@ public class CategoryServiceImpl implements CategoryService {
         if (!categoryRepository.existsById(id)) {
             throw new NotFoundException("События с id=" + id + " не существует");
         }
-        //Добавить проверку связи категории с событиями
+        if (eventRepository.existsByCategoryId(id)) {
+            throw new AlreadyExistsException("Нельзя удалить категорию к которой есть привязанные события");
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -67,7 +68,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDto getCategoryById(long id) {
         Category category = categoryRepository.findById(id)
-                .orElseThrow(()-> new NotFoundException("Категории с id=" + id + " не найдено"));
+                .orElseThrow(() -> new NotFoundException("Категории с id=" + id + " не найдено"));
         return CategoryMapper.mapToCategoryDto(category);
     }
 }
